+++ src/saslite/__init__.py
@@ -1,6 +1,6 @@
 """SASLite - Lightweight local SAS language interpreter."""
 
-__version__ = "0.4.1+custom.6"
+__version__ = "0.4.1+custom.7"
 
 from saslite.api.facade import SasInterpreter